// src/domain/specialization/doctor-specialty.enum.ts
export enum DoctorSpecialty {
  CARDIOLOGY = 'CARDIOLOGY',
  DERMATOLOGY = 'DERMATOLOGY',
  ENDOCRINOLOGY = 'ENDOCRINOLOGY',
  GASTROENTEROLOGY = 'GASTROENTEROLOGY',
  HEMATOLOGY = 'HEMATOLOGY',
  NEUROLOGY = 'NEUROLOGY',
  OBSTETRICS_GYNECOLOGY = 'OBSTETRICS & GYNECOLOGY',
  ONCOLOGY = 'ONCOLOGY',
  OPHTHALMOLOGY = 'OPHTHALMOLOGY',
  ORTHOPEDICS = 'ORTHOPEDICS',
  OTOLARYNGOLOGY = 'OTOLARYNGOLOGY (ENT)',
  PEDIATRICS = 'PEDIATRICS',
  PSYCHIATRY = 'PSYCHIATRY',
  RADIOLOGY = 'RADIOLOGY',
  UROLOGY = 'UROLOGY',
}